refactor: GameResult 추상화 후 RacingGameResult를 구현
