+++ src/main/java/racingcar/domain/GameResult.java
@@ -1,20 +1,7 @@
 package racingcar.domain;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
+public interface GameResult {
+    void addGameStatus(GameStatus status);
 
-public class GameResult {
-    private static final String GAME_STATUS_SEPERATOR = "\n\n";
-    private final List<GameStatus> gameStatusBucket = new ArrayList<>();
-
-    public void addGameStatus(GameStatus status) {
-        gameStatusBucket.add(status);
-    }
-
-    public String getGameResult() {
-        return gameStatusBucket.stream()
-                .map(GameStatus::getStatus)
-                .collect(Collectors.joining(GAME_STATUS_SEPERATOR));
-    }
+    String getGameResult();
 }
